+++ src/main/java/com/example/forummanagementsystem/repositories/UserRepositoryImpl.java
@@ -32,7 +32,7 @@ public class UserRepositoryImpl implements UserRepository {
     }
 
     @Override
-    public User get(int id) {
+    public User get(Long id) {
         try (Session session = sessionFactory.openSession()) {
             User user = session.get(User.class, id);
             if (user == null) {
