+++ src/main/java/com/example/forummanagementsystem/services/UserServiceImpl.java
@@ -25,7 +25,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User get(int id) {
+    public User get(Long id) {
         return userRepository.get(id);
     }
